Need uint64_t for sampling rate type.

+++ plugins/collectors/hwc/HWCCollector.cxx
@@ -77,7 +77,7 @@ HWCCollector::HWCCollector() :
     // Declare our parameters
     declareParameter(Metadata("sampling_rate", "Sampling Threshhold",
                               "Sampling threshhold in hardware counter events.",
-                              typeid(unsigned)));
+                              typeid(uint64_t)));
 
     declareParameter(Metadata("event", "Hardware Counter Event",
                            "HWC event.",
@@ -120,9 +120,9 @@ Blob HWCCollector::getDefaultParameterValues() const
 
 #if defined(linux)
     if (hw_info) {
-	parameters.sampling_rate = (unsigned) hw_info->mhz*10000*2;
+	parameters.sampling_rate = (uint64_t) hw_info->mhz*10000*2;
     } else {
-	parameters.sampling_rate = (unsigned) THRESHOLD*2;
+	parameters.sampling_rate = (uint64_t) THRESHOLD*2;
     }
 #else
     parameters.sampling_rate = THRESHOLD*2;
@@ -158,7 +158,7 @@ void HWCCollector::getParameterValue(const std::string& parameter,
 
     // Handle the "sampling_rate" parameter
     if(parameter == "sampling_rate") {
-        unsigned* value = reinterpret_cast<unsigned*>(ptr);
+        uint64_t* value = reinterpret_cast<uint64_t*>(ptr);
         *value = parameters.sampling_rate;
     }
 
@@ -194,7 +194,7 @@ void HWCCollector::setParameterValue(const std::string& parameter,
     
     // Handle the "sampling_rate" parameter
     if(parameter == "sampling_rate") {
-        const unsigned* value = reinterpret_cast<const unsigned*>(ptr);
+        const uint64_t* value = reinterpret_cast<const uint64_t*>(ptr);
         parameters.sampling_rate = *value;
 	std::ostringstream rate;
 	rate << parameters.sampling_rate;
